Оптимизация меню #33
